requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
dj-database-url==2.3.0
Django==5.2
django-auditlog==3.0.0
django-simple-history==3.8.0
django_celery_results==2.6.0
et_xmlfile==2.0.0
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.3
msgpack==1.1.0
numpy==2.2.4
openpyxl==3.1.5
openrouteservice==2.3.3
packaging==25.0
pandas==2.2.3
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
psycopg2-binary>=2.9
pytz==2025.2
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
setuptools==80.0.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2